Branch 134070499

+++ tensorflow/core/lib/strings/str_util.h
@@ -162,7 +162,7 @@ inline std::vector<string> Split(StringPiece text, char delim) {
 template <typename Predicate>
 std::vector<string> Split(StringPiece text, char delim, Predicate p) {
   std::vector<string> result;
-  int token_start = 0;
+  size_t token_start = 0;
   if (!text.empty()) {
     for (size_t i = 0; i < text.size() + 1; i++) {
       if ((i == text.size()) || (text[i] == delim)) {
